+++ src/parser/character/spellCasting.js
@@ -183,4 +183,5 @@ DDBCharacter.prototype._generateMaxPreparedSpells = function _generateMaxPrepare
 
   this.raw.character.system.details.maxPreparedSpells = max;
   setProperty(this.raw.character, "flags.tidy5e-sheet.maxPreparedSpells", max);
+  setProperty(this.raw.character, "flags.tidy5e-sheet-kgar.maxPreparedSpells", max);
 };